requirements.txt
fastapi[standard]
requests
boto3
dotenv
openai
pydantic-settings